refactor: cleaning up PR
