+++ blog/scripts/scripts.js
@@ -1135,206 +1135,13 @@ function parseExperimentConfig(json) {
 }
 
 /**
- * Gets experiment config from the manifest and transforms it to more easily
- * consumable structure.
- *
- * the manifest consists of two sheets "settings" and "experiences", by default
- *
- * "settings" is applicable to the entire test and contains information
- * like "Audience", "Status" or "Blocks".
- *
- * "experience" hosts the experiences in rows, consisting of:
- * a "Percentage Split", "Label" and a set of "Links".
- *
- *
- * @param {string} experimentId
- * @param {object} cfg
- * @returns {object} containing the experiment manifest
- */
- export async function getExperimentConfig(experimentId, instantExperiment) {
-  if (instantExperiment) {
-    const config = {
-      experimentName: `Instant Experiment: ${experimentId}`,
-      audience: '',
-      status: 'Active',
-      id: experimentId,
-      variants: {},
-      variantNames: [],
-    };
-
-    const pages = instantExperiment.split(',').map((p) => new URL(p.trim()).pathname);
-    const evenSplit = 1 / (pages.length + 1);
-
-    config.variantNames.push('control');
-    config.variants.control = {
-      percentageSplit: '',
-      pages: [window.location.pathname],
-      blocks: [],
-      label: 'Control',
-    };
-
-    pages.forEach((page, i) => {
-      const vname = `challenger-${i + 1}`;
-      config.variantNames.push(vname);
-      config.variants[vname] = {
-        percentageSplit: `${evenSplit}`,
-        pages: [page],
-        label: `Challenger ${i + 1}`,
-      };
-    });
-
-    return (config);
-  }
-
-  const path = `${cfg.basePath}/${experimentId}/${cfg.configFile}`;
-  try {
-    const resp = await fetch(path);
-    if (!resp.ok) {
-      console.log('error loading experiment config:', resp);
-      return null;
-    }
-    const json = await resp.json();
-    const config = cfg.parser ? cfg.parser(json) : parseExperimentConfig(json);
-    config.id = experimentId;
-    config.manifest = path;
-    config.basePath = `${cfg.basePath}/${experimentId}`;
-    return config;
-  } catch (e) {
-    console.log(`error loading experiment manifest: ${path}`, e);
-  }
-  return null;
-}
-
-/**
- * this is an extensible stub to take on audience mappings
- * @param {string} audience
- * @return {boolean} is member of this audience
- */
-function isValidAudience(audience) {
-  if (audience === 'mobile') {
-    return window.innerWidth < 600;
-  }
-  if (audience === 'desktop') {
-    return window.innerWidth >= 600;
-  }
-  return true;
-}
-
-function getDecisionPolicy(config) {
-  const decisionPolicy = {
-    id: 'content-experimentation-policy',
-    rootDecisionNodeId: 'n1',
-    decisionNodes: [{
-      id: 'n1',
-      type: 'EXPERIMENTATION',
-      experiment: {
-        id: config.id,
-        identityNamespace: 'ECID',
-        randomizationUnit: 'DEVICE',
-        treatments: Object.entries(config.variants).map(([key, props]) => ({
-          id: key,
-          allocationPercentage: props.percentageSplit
-            ? parseFloat(props.percentageSplit) * 100
-            : 100 - Object.values(config.variants).reduce((result, variant) => {
-              result -= parseFloat(variant.percentageSplit || 0) * 100;
-              return result;
-            }, 100),
-        })),
-      },
-    }],
-  };
-  return decisionPolicy;
-}
-
-/**
- * Replaces element with content from path
- * @param {string} path
- * @param {HTMLElement} element
- * @param {boolean} isBlock
- */
-async function replaceInner(path, element, isBlock = false) {
-  const plainPath = `${path}.plain.html`;
-  try {
-    const resp = await fetch(plainPath);
-    if (!resp.ok) {
-      console.log('error loading experiment content:', resp);
-      return null;
-    }
-    const html = await resp.text();
-    if (isBlock) {
-      const div = document.createElement('div');
-      div.innerHTML = html;
-      element.replaceWith(div.children[0].children[0]);
-    } else {
-      element.innerHTML = html;
-    }
-  } catch (e) {
-    console.log(`error loading experiment content: ${plainPath}`, e);
-  }
-  return null;
-}
-
-async function runExperiment(experiment, instantExperiment) {
-  const ued = import ('./ued.js');
-  const usp = new URLSearchParams(window.location.search);
-  const [forcedExperiment, forcedVariant] = usp.has('experiment') ? usp.get('experiment').split('/') : [];
-
-  const experimentConfig = await getExperimentConfig(experiment, instantExperiment);
-  console.debug(experimentConfig);
-  if (!experimentConfig || (toCamelCase(experimentConfig.status) !== 'active' && !forcedExperiment)) {
-    return;
-  }
-
-  experimentConfig.run = forcedExperiment
-    || isValidAudience(toClassName(experimentConfig.audience));
-  window.hlx = window.hlx || {};
-  window.hlx.experiment = experimentConfig;
-  console.debug('run', experimentConfig.run, experimentConfig.audience);
-  if (!experimentConfig.run) {
-    return;
-  }
-
-  if (forcedVariant && experimentConfig.variantNames.includes(forcedVariant)) {
-    experimentConfig.selectedVariant = forcedVariant;
-  } else {
-    const { evaluateDecisionPolicy } = await ued;
-    const decision = evaluateDecisionPolicy(getDecisionPolicy(experimentConfig), {});
-    experimentConfig.selectedVariant = decision.items[0].id;
-  }
-
-  sampleRUM('experiment', { source: experimentConfig.id, target: experimentConfig.selectedVariant });
-  console.debug(`running experiment (${window.hlx.experiment.id}) -> ${window.hlx.experiment.selectedVariant}`);
-
-  if (experimentConfig.selectedVariant === experimentConfig.variantNames[0]) {
-    return;
-  }
-
-  const currentPath = window.location.pathname;
-  const { pages } = experimentConfig.variants[experimentConfig.selectedVariant];
-  if (!pages.length) {
-    return;
-  }
-
-  const control = experimentConfig.variants[experimentConfig.variantNames[0]];
-  const index = control.pages.indexOf(currentPath);
-  if (index < 0 || pages[index] === currentPath) {
-    return;
-  }
-
-  // Fullpage content experiment
-  const main = document.querySelector('main');
-  await replaceInner(pages[0], main);
-  main.classList.add(experimentConfig.id);
-  main.classList.add(experimentConfig.selectedVariant);
-}
-
-/**
  * loads everything needed to get to LCP.
  */
 async function loadEager(doc) {
   const experiment = getMetadata('experiment');
   const instantExperiment = getMetadata('instant-experiment');
   if (instantExperiment || experiment) {
+    const { runExperiment } = await import('./experimentation.js');
     await runExperiment(experiment, instantExperiment);
   }
